Face-Mask-Detection/Mask detection using MTCNN/requirements.txt
certifi==2020.4.5.1
flake8==3.7.9
future==0.18.2
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
Pillow==7.1.2
pycodestyle==2.5.0
pyflakes==2.1.1
scipy==1.4.1
six==1.14.0
torch==1.5.0
torch-mtcnn==0.0.7
torchvision==0.6.0a0+82fd1c8
